Add brutefir_config option
The idea behind this is to add a new flag to snapclient,
--brutefir_config, which specifies the path to the brutefir config file.
If this is not provided, brutefir isn't used to filter the data.

If it is provided, we'll pass this to brutefir as the config to use.
We'll also make sure brutefir's working directory is the same directory
as the config file, which will allow using relative coefficient paths in
the configuration file.